Refactor to src layout and improve tooling
