+++ .env.example
@@ -1,0 +1,2 @@
+# Example environment configuration
+LOG_LEVEL=INFO
