+++ requirements-dev.txt
@@ -3,3 +3,5 @@ flake8
 pytest
 pytest-cov
 pre-commit
+ruff
+pip-audit
