+++ src/wcr_data_api/__init__.py
@@ -1,0 +1,1 @@
+from .main import app
